+++ src/mwparserfromhell/parser/ctokenizer/avl_tree.c
@@ -19,7 +19,7 @@
  */
 
 #define false 0
-#define true 1
+#define true  1
 
 typedef int bool;